assignment-3/pneumonia-detection/src/api/requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
python-multipart>=0.0.5
pillow>=8.3.1
torch>=1.9.0
torchvision>=0.10.0
python-dotenv>=0.19.0
pydantic>=1.8.2
pydantic_settings>=1.8.2
redis>=4.0.0
aioredis>=2.0.0
tenacity>=8.0.0